+++ src/main.rs
@@ -8,14 +8,14 @@ use crate::api::display::Point;
 use app::App;
 use crossterm::terminal::size;
 
-fn get_screen_size() -> Point<u32> {
+fn get_screen_size() -> Point<i64> {
     let (rows, columns) = size().unwrap();
 
-    Point::new(rows as u32, columns as u32)
+    Point::new(rows as i64, columns as i64)
 }
 
 fn main() {
-    // let dimensions: &Point<u32> = &Point {
+    // let dimensions: &Point<i64> = &Point {
     //     width: 60,
     //     height: 40,
     // };
